Move beep to it's own module
