+++ src/core/ht.c
@@ -17,7 +17,7 @@
 #include "bmi270/accel_gyro.h"
 #include "core/settings.h"
 #include "driver/dm6302.h"
-#include "driver/gpio.h"
+#include "driver/beep.h"
 #include "driver/hardware.h"
 #include "driver/oled.h"
 #include "ui/page_common.h"
