+++ src/driver/gpio.h
@@ -2,11 +2,6 @@
 
 #include <stdbool.h>
 
-#define BEEP_SHORT 50
-#define BEEP_LONG 200
-
-#define beep() beep_dur(BEEP_SHORT)
-
 void gpio_init();
 void gpio_open(int port_num);
 void gpio_set(int port_num, bool val);
