+++ src/ui/page_version.c
@@ -15,7 +15,7 @@
 #include "driver/dm5680.h"
 #include "driver/esp32.h"
 #include "driver/fans.h"
-#include "driver/gpio.h"
+#include "driver/beep.h"
 #include "driver/i2c.h"
 #include "driver/uart.h"
 #include "ui/page_common.h"
