+++ src/ui/ui_statusbar.c
@@ -8,7 +8,7 @@
 #include "core/common.hh"
 #include "core/osd.h"
 #include "core/settings.h"
-#include "driver/gpio.h"
+#include "driver/beep.h"
 #include "ui/page_common.h"
 #include "ui/page_playback.h"
 #include "ui/ui_style.h"
